renamed charm library to fit new convention
